dealer-cards in component
